+++ main.go
@@ -3,7 +3,6 @@ package main
 import (
 	"fmt"
 	"os"
-	"os/exec"
 
 	"github.com/spf13/cobra"
 )
@@ -18,12 +17,18 @@ var rootCmd = &cobra.Command{
 	},
 }
 
-func main() {
-	// Ensure we have tar/gzip/curl installed.
-	CheckForTool("tar")
-	CheckForTool("gzip")
-	CheckForTool("curl")
+func init() {
+	importCmd.Flags().String("path", "/app/backup.tar.gz", "Path of tar to import")
+	importCmd.Flags().String("http-method", "GET", "HTTP method to use")
+	rootCmd.AddCommand(importCmd)
+	exportCmd.Flags().String("path", "/app/backup.tar.gz", "Path of tar to export") // Path or URL where the backup will be stored.
+	exportCmd.Flags().String("http-method", "PUT", "HTTP method to use")
+	rootCmd.AddCommand(exportCmd)
+	rootCmd.AddCommand(sizeCmd)
+	rootCmd.AddCommand(destroyCmd)
+}
 
+func main() {
 	// Ensure /data directory exists.
 	if _, err := os.Stat(pvDirectory); os.IsNotExist(err) {
 		fmt.Printf("Directory %s does not exist", pvDirectory)
@@ -37,9 +42,71 @@ func main() {
 	rootCmd.Execute()
 }
 
-func CheckForTool(tool string) {
-	_, err := exec.LookPath(tool)
-	if err != nil {
-		panic(fmt.Sprintf("Could not find %s in PATH", tool))
-	}
+var importCmd = &cobra.Command{
+	Use: "import",
+	Run: func(cmd *cobra.Command, args []string) {
+		// If the path is started with http , assume it is a URL and download the file.
+		path := cmd.Flag("path").Value.String()
+		var err error
+		if len(path) > 4 && path[:4] == "http" {
+			method := cmd.Flag("http-method").Value.String()
+			err = RestoreFromURL(path, method)
+		} else {
+			err = RestoreFromFile(path)
+		}
+		if err != nil {
+			PrintError(err.Error())
+		}
+		PrintData("Data has been imported")
+	},
+}
+
+var exportCmd = &cobra.Command{
+	Use: "export",
+	Run: func(cmd *cobra.Command, args []string) {
+		path := cmd.Flag("path").Value.String()
+		var err error
+		// If the path is started with http , assume it is a URL and upload the file.
+		if len(path) > 4 && path[:4] == "http" {
+			method := cmd.Flag("http-method").Value.String()
+			err = BackupToURL(path, method)
+		} else {
+			err = BackupToFile(path)
+		}
+		if err != nil {
+			PrintError(err.Error())
+		}
+		PrintData("Data has been exported")
+	},
+}
+
+var sizeCmd = &cobra.Command{
+	Use: "size",
+	Run: func(cmd *cobra.Command, args []string) {
+		size, err := FetchFileSize(pvDirectory)
+		if err != nil {
+			size = 0
+		}
+		// Write size to /app/size.txt for backward compatibility.
+		file := "/app/size.txt"
+		f, err := os.Create(file)
+		if err != nil {
+			PrintError(err.Error())
+		}
+		defer f.Close()
+		_, err = f.WriteString(fmt.Sprintf("%d", size))
+		if err != nil {
+			PrintError(err.Error())
+		}
+		PrintData(size)
+	},
+}
+
+var destroyCmd = &cobra.Command{
+	Use: "destroy",
+	Run: func(cmd *cobra.Command, args []string) {
+		// Destroy the backup.
+		_ = os.RemoveAll(pvDirectory)
+		PrintData("All data has been destroyed")
+	},
 }
